pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-avatar-video-generation"
version = "1.0.0"
description = "Local AI avatar video generation pipeline and Gradio dashboard."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate",
    "diffusers>=0.25.0",
    "faster-whisper>=1.0.0",
    "gradio>=4.0.0",
    "insightface>=0.7.0",
    "kokoro>=0.9.4",
    "kornia>=0.7.0",
    "loguru",
    "mlx-audio>=0.3.0",
    "numpy",
    "pillow",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml",
    "requests",
    "soundfile",
    "tqdm",
    "transformers>=4.36.0",
]

[project.optional-dependencies]
japanese = [
    "fugashi[unidic-lite]>=1.3.2",
    "mojimoji>=0.0.13",
    "pykakasi>=2.3.0",
    "pyopenjtalk>=0.4.1",
]

[project.scripts]
avatar-dashboard = "scripts.run_dashboard:main"
avatar-pipeline = "scripts.run_pipeline:main"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["avatarpipeline*", "app", "scripts"]
exclude = ["tests*", "data*", "doc*", "install*", "tools*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
markers = [
    "slow: tests that run model inference or large downloads",
]
